add trash product list

+++ app/controllers/io_store/carts_controller.rb
@@ -2,16 +2,15 @@ module IoStore
   class CartsController < IoStore::ApplicationController
 
 
-    before_filter :authenticate_user!
+    before_filter :authenticate_user!, :pre_load
+
+    def pre_load
+      extract_cart if @cart.nil?
+    end
 
 
     def index
-      if session[:shopping_product_id]
-        session.delete(:shopping_product_id)
-      end
-
-      extract_cart if @cart.nil?
-      
+      session.delete(:shopping_product_id) if session[:shopping_product_id]
     end
 
     def add
